Refactor: Change updateLocalStorage function to update localStorage directly and Change name
